+++ test/unit/fuzz-tools.test.js
@@ -395,6 +395,7 @@ import * as geisslerCarpalV415 from '../../lib/geissler-carpal-v415.js';
 import * as russeScaphoidV416 from '../../lib/russe-scaphoid-v416.js';
 import * as wasselThumbV417 from '../../lib/wassel-thumb-v417.js';
 import * as milchCondyleV418 from '../../lib/milch-condyle-v418.js';
+import * as cottonMyerV419 from '../../lib/cotton-myer-v419.js';
 import * as tbTesting from '../../lib/tb-testing.js';
 
 const MODULES = {
@@ -735,6 +736,7 @@ const MODULES = {
   'russe-scaphoid-v416.js': russeScaphoidV416,
   'wassel-thumb-v417.js': wasselThumbV417,
   'milch-condyle-v418.js': milchCondyleV418,
+  'cotton-myer-v419.js': cottonMyerV419,
   'tb-testing.js': tbTesting,
 };